Edit migration

+++ database/migrations/2021_05_08_222047_create_magnitudes_table.php
@@ -14,9 +14,10 @@ class CreateMagnitudesTable extends Migration
     public function up()
     {
         Schema::create('magnitudes', function (Blueprint $table) {
+
             $table->id();
             $table->string('name',10);
-            $table->float('size');
+            $table->decimal('size', 10, 10);
 
             $table->timestamps();
         });
